@@ -8,72 +8,67 @@
     "start": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test",
-    "test:ci": "ng test --browsers=Headless --no-watch",
+    "test": "jest",
+    "test:coverage": "jest --coverage",
+    "test:ci": "jest -ci",
     "lint": "ng lint",
     "eje": "copy customTheme customTheme2"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^15.0.0",
-    "@angular/core": "^15.0.0",
-    "@angular/forms": "^15.0.0",
-    "@angular/platform-browser": "^15.0.0",
-    "@angular/platform-browser-dynamic": "^15.0.0",
-    "@angular/router": "^15.0.0",
-    "@capacitor/android": "4.7.3",
-    "@capacitor/app": "4.1.1",
-    "@capacitor/core": "4.7.3",
-    "@capacitor/haptics": "4.1.0",
-    "@capacitor/keyboard": "4.1.1",
-    "@capacitor/status-bar": "4.1.1",
-    "@ionic/angular": "^7.0.0",
-    "@ionic/storage-angular": "^3.0.6",
-    "@ngx-translate/core": "^14.0.0",
-    "@ngx-translate/http-loader": "^7.0.0",
-    "@zxing/browser": "^0.1.1",
-    "@zxing/library": "^0.19.2",
-    "@zxing/ngx-scanner": "^3.8.1",
-    "angular-auth-oidc-client": "^16.0.1",
-    "angularx-qrcode": "^15.0.1",
-    "ionicons": "^7.0.0",
-    "ngx-indexed-db": "^16.0.0",
-    "rxjs": "~7.5.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "@angular/common": "^18.2.2",
+    "@angular/core": "^18.2.2",
+    "@angular/forms": "^18.2.2",
+    "@angular/platform-browser": "^18.2.2",
+    "@angular/platform-browser-dynamic": "^18.2.2",
+    "@angular/router": "^18.2.2",
+    "@capacitor/android": "6.1.2",
+    "@capacitor/app": "6.0.1",
+    "@capacitor/core": "6.1.2",
+    "@capacitor/haptics": "6.0.1",
+    "@capacitor/keyboard": "6.0.2",
+    "@capacitor/status-bar": "6.0.1",
+    "@ionic/angular": "^8.2.7",
+    "@ionic/storage-angular": "^4.0.0",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
+    "@zxing/browser": "^0.1.5",
+    "@zxing/library": "^0.21.3",
+    "@zxing/ngx-scanner": "^18.0.1",
+    "angular-auth-oidc-client": "^18.0.1",
+    "angularx-qrcode": "^18.0.1",
+    "ionicons": "^7.4.0",
+    "ngx-indexed-db": "^19.0.0",
+    "rxjs": "~7.8.1",
+    "tslib": "^2.7.0",
+    "wallet-ui": "file:",
+    "zone.js": "~0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.0.0",
-    "@angular-eslint/builder": "15.2.1",
-    "@angular-eslint/eslint-plugin": "15.2.1",
-    "@angular-eslint/eslint-plugin-template": "15.2.1",
-    "@angular-eslint/schematics": "15.2.1",
-    "@angular-eslint/template-parser": "15.2.1",
-    "@angular/cli": "^15.0.0",
-    "@angular/compiler": "^15.0.0",
-    "@angular/compiler-cli": "^15.0.0",
-    "@angular/language-service": "^15.0.0",
-    "@babel/plugin-proposal-decorators": "^7.23.9",
-    "@capacitor/cli": "4.7.3",
-    "@ionic/angular-toolkit": "^9.0.0",
-    "@types/jasmine": "~4.0.0",
-    "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "5.48.2",
-    "@typescript-eslint/parser": "5.48.2",
-    "eslint": "^8.33.0",
-    "jasmine-core": "~4.3.0",
-    "jasmine-spec-reporter": "~5.0.0",
+    "@angular-devkit/build-angular": "^18.2.2",
+    "@angular-eslint/builder": "18.3.0",
+    "@angular-eslint/eslint-plugin": "18.3.0",
+    "@angular-eslint/eslint-plugin-template": "18.3.0",
+    "@angular-eslint/schematics": "18.3.0",
+    "@angular-eslint/template-parser": "18.3.0",
+    "@angular/cli": "^18.2.2",
+    "@angular/compiler": "^18.2.2",
+    "@angular/compiler-cli": "^18.2.2",
+    "@angular/language-service": "^18.2.2",
+    "@babel/plugin-proposal-decorators": "^7.24.7",
+    "@capacitor/cli": "6.1.2",
+    "@ionic/angular-toolkit": "^12.1.1",
+    "@types/jest": "^29.5.13",
+    "@types/node": "^22.5.3",
+    "@typescript-eslint/eslint-plugin": "8.4.0",
+    "@typescript-eslint/parser": "8.4.0",
+    "eslint": "^9.9.1",
+    "jasmine-spec-reporter": "~7.0.0",
     "jest": "^29.7.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "karma-sonarqube-reporter": "^1.4.0",
-    "ng-mocks": "^14.12.1",
-    "ts-node": "~8.3.0",
-    "typescript": "~4.8.4"
+    "jest-preset-angular": "^14.2.4",
+    "ng-mocks": "^14.13.1",
+    "ts-node": "~10.9.2",
+    "typescript": "~5.5.4"
   },
   "description": "An Ionic project"
 }
