@@ -1,9 +1,10 @@
 import { TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
-import { HTTP_INTERCEPTORS, HttpClient, HttpErrorResponse } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, HttpClient } from '@angular/common/http';
 import { ToastServiceHandler } from '../services/toast.service';
 import { HttpErrorInterceptor } from './error-handler.interceptor';
-/*
+import { environment } from 'src/environments/environment';
+
 class MockToastServiceHandler {
   showErrorAlert(message: string) {
   }
@@ -39,23 +40,25 @@ describe('HttpErrorInterceptor with HttpClient', () => {
   afterEach(() => {
     httpMock.verify();
   });
+
   it('should log and show a toast on a 404 Not Found response', () => {
-    const spy = spyOn(mockToastServiceHandler, 'showErrorAlert');
-    const consoleErrorSpy = spyOn(console, 'error');
+    const spy = jest.spyOn(mockToastServiceHandler, 'showErrorAlert');
+    const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
 
     httpClient.get('/test404').subscribe({
       error: (error) => {
         expect(spy).toHaveBeenCalledWith('Resource not found message from backend');
-        expect(consoleErrorSpy).toHaveBeenCalledWith('Resource not found:', jasmine.any(String));
+        expect(consoleErrorSpy).toHaveBeenCalledWith('Resource not found:', expect.any(String));
       }
     });
 
     const req = httpMock.expectOne('/test404');
     req.flush({message: 'Resource not found message from backend'}, { status: 404, statusText: 'Not Found' });
   });
+
   it('should show error toast on 422 Unprocessable Entity response', () => {
     const expectedMessage = 'Unprocessable Entity';
-    const spy = spyOn(mockToastServiceHandler, 'showErrorAlert');
+    const spy = jest.spyOn(mockToastServiceHandler, 'showErrorAlert');
 
     httpClient.get('/test422').subscribe({
       error: (error) => {
@@ -69,7 +72,7 @@ describe('HttpErrorInterceptor with HttpClient', () => {
 
   it('should show error toast on 500 Internal Server Error response', () => {
     const expectedMessage = 'Internal Server Error';
-    const spy = spyOn(mockToastServiceHandler, 'showErrorAlert');
+    const spy = jest.spyOn(mockToastServiceHandler, 'showErrorAlert');
 
     httpClient.get('/test500').subscribe({
       error: (error) => {
@@ -81,40 +84,48 @@ describe('HttpErrorInterceptor with HttpClient', () => {
     req.flush({ message: expectedMessage }, { status: 500, statusText: 'Internal Server Error' });
   });
 
+  it('should log and show a toast on a 401 Unauthorized response', () => {
+    const spy = jest.spyOn(mockToastServiceHandler, 'showErrorAlert');
+    const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
 
-it('should log and show a toast on a 401 Unauthorized response', () => {
-  const spy = spyOn(mockToastServiceHandler, 'showErrorAlert');
-  const consoleErrorSpy = spyOn(console, 'error');
+    httpClient.get('/test401').subscribe({
+      error: (error) => {
+        expect(spy).toHaveBeenCalledWith('Unauthorized error message from backend');
+        expect(consoleErrorSpy).toHaveBeenCalledWith('Unauthorized:', expect.any(String));
+      }
+    });
 
-  httpClient.get('/test401').subscribe({
-    error: (error) => {
-      expect(spy).toHaveBeenCalledWith('Unauthorized error message from backend');
-      expect(consoleErrorSpy).toHaveBeenCalledWith('Unauthorized:', jasmine.any(String));
-    }
+    const req = httpMock.expectOne('/test401');
+    req.flush({message: 'Unauthorized error message from backend'}, { status: 401, statusText: 'Unauthorized' });
   });
 
-  const req = httpMock.expectOne('/test401');
-  req.flush({message: 'Unauthorized error message from backend'}, { status: 401, statusText: 'Unauthorized' });
-});
+  it('should log and show a toast on a generic HTTP error response', () => {
+    const errorMessage = 'An error occurred';
+    const spy = jest.spyOn(mockToastServiceHandler, 'showErrorAlert');
+    const consoleErrorSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
 
-it('should log and show a toast on a generic HTTP error response', () => {
-  const errorMessage = 'An error occurred';
-  const spy = spyOn(mockToastServiceHandler, 'showErrorAlert');
-  const consoleErrorSpy = spyOn(console, 'error');
+    httpClient.get('/testError').subscribe({
+      error: (error) => {
+        expect(spy).toHaveBeenCalledWith(errorMessage);
+        expect(consoleErrorSpy).toHaveBeenCalledWith('An HTTP error occurred:', expect.any(String));
+      }
+    });
 
-  httpClient.get('/testError').subscribe({
-    error: (error) => {
-      expect(spy).toHaveBeenCalledWith(errorMessage);
-      expect(consoleErrorSpy).toHaveBeenCalledWith('An HTTP error occurred:', jasmine.any(String));
-    }
+    const req = httpMock.expectOne('/testError');
+    req.flush({message: errorMessage}, { status: 500, statusText: 'Internal Server Error' });
   });
 
-  const req = httpMock.expectOne('/testError');
-  req.flush({message: errorMessage}, { status: 500, statusText: 'Internal Server Error' });
+  it('it should print the correct message if a request to process the QR is made', ()=>{
+    const errorMessage = 'There was a problem processing the QR. It might be invalid or already have been used';
+    const spy = jest.spyOn(mockToastServiceHandler, 'showErrorAlert');
+
+    httpClient.get('/' + environment.server_uri.execute_content_uri).subscribe({
+      error: (error) => {
+        expect(spy).toHaveBeenCalledWith('errorMessage');
+      }
+    });
+
+    const req = httpMock.expectOne('/' + environment.server_uri.execute_content_uri);
+    req.flush({message: 'Random error message'}, { status: 500, statusText: 'AnyText' });
+  });
 });
-
-
-
-});
-
-*/
