@@ -2,8 +2,8 @@ import { ComponentFixture, TestBed, fakeAsync, tick, waitForAsync } from '@angul
 import { AlertController, IonicModule } from '@ionic/angular';
 import { TranslateModule } from '@ngx-translate/core';
 import { RouterTestingModule } from '@angular/router/testing';
-import { ActivatedRoute, Router } from '@angular/router';
-import { of } from 'rxjs';
+import { ActivatedRoute, NavigationEnd, Router } from '@angular/router';
+import { of, throwError, Subject } from 'rxjs';
 import { CredentialsPage } from './credentials.page';
 import { WalletService } from 'src/app/services/wallet.service';
 import { WebsocketService } from 'src/app/services/websocket.service';
@@ -11,27 +11,53 @@ import { DataService } from 'src/app/services/data.service';
 import { AuthenticationService } from 'src/app/services/authentication.service';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { CredentialStatus, VerifiableCredential } from 'src/app/interfaces/verifiable-credential';
+import { Storage } from '@ionic/storage-angular';
+
+class MockRouter {
+  public events = new Subject<any>();
+  public navigate = (route:string|string[], opt?:{})=>'';
+}
+
+const writeText = jest.fn()
+Object.assign(navigator, {
+  clipboard: {
+    writeText,
+  },
+});
 
 describe('CredentialsPage', () => {
   let component: CredentialsPage;
   let fixture: ComponentFixture<CredentialsPage>;
-  let walletServiceSpy: jasmine.SpyObj<WalletService>;
-  let websocketServiceSpy: jasmine.SpyObj<WebsocketService>;
+  let walletServiceSpy: jest.Mocked<WalletService>;
+  let websocketServiceSpy: jest.Mocked<WebsocketService>;
   let httpTestingController: HttpTestingController;
-  let router: Router;
+  let mockRouter: MockRouter;
 
   const TIME_IN_MS = 10000;
 
   beforeEach(waitForAsync(() => {
-    walletServiceSpy = jasmine.createSpyObj('WalletService', ['getAllVCs', 'requestCredential', 'deleteVC', 'executeContent']);
-    websocketServiceSpy = jasmine.createSpyObj('WebsocketService', ['connect']);
-    const dataServiceSpyObj = jasmine.createSpyObj('DataService', ['listenDid']);
-    const authServiceSpyObj = jasmine.createSpyObj('AuthenticationService', ['getName']);
-    walletServiceSpy.requestCredential.and.returnValue(of({} as any));
 
-    dataServiceSpyObj.listenDid.and.returnValue(of('someDidValue'));
-    walletServiceSpy.executeContent.and.returnValue(of({} as any));
-    walletServiceSpy.requestCredential.and.returnValue(of({} as any));
+    walletServiceSpy = {
+      getAllVCs: jest.fn().mockReturnValue(of([])),
+      requestCredential: jest.fn().mockReturnValue(of({} as any)),
+      deleteVC: jest.fn(),
+      executeContent: jest.fn().mockReturnValue(of({} as any))
+    } as unknown as jest.Mocked<WalletService>;
+
+    websocketServiceSpy = {
+      connect: jest.fn(),
+      closeConnection: jest.fn()
+    } as unknown as jest.Mocked<WebsocketService>;
+
+    const dataServiceSpyObj = {
+      listenDid: jest.fn().mockReturnValue(of('someDidValue'))
+    } as unknown as jest.Mocked<DataService>;
+
+    const authServiceSpyObj = {
+      getName: jest.fn()
+    } as unknown as jest.Mocked<AuthenticationService>;
+
+    mockRouter = new MockRouter();
 
     TestBed.configureTestingModule({
       imports: [
@@ -41,6 +67,8 @@ describe('CredentialsPage', () => {
         HttpClientTestingModule
       ],
       providers: [
+        Storage,
+        { provide: Router, useValue: mockRouter},
         { provide: WalletService, useValue: walletServiceSpy },
         { provide: WebsocketService, useValue: websocketServiceSpy },
         { provide: DataService, useValue: dataServiceSpyObj },
@@ -49,6 +77,11 @@ describe('CredentialsPage', () => {
           provide: ActivatedRoute,
           useValue: {
             queryParams: of({ credentialOfferUri: 'mockCredentialOfferUri' }),
+            snapshot: {
+              routeConfig: {
+                path: 'credentials'
+              }
+            }
           },
         },
       ],
@@ -57,9 +90,7 @@ describe('CredentialsPage', () => {
     httpTestingController = TestBed.inject(HttpTestingController);
     fixture = TestBed.createComponent(CredentialsPage);
     component = fixture.componentInstance;
-    router = TestBed.inject(Router);
   }));
-
 
   it('should create', () => {
     expect(component).toBeTruthy();
@@ -67,45 +98,47 @@ describe('CredentialsPage', () => {
 
   it('should enable scan mode when scan is called', () => {
     component.scan();
-    expect(component.toggleScan).toBeTrue();
-    expect(component.show_qr).toBeTrue();
-    expect(component.ebsiFlag).toBeFalse();
+    expect(component.toggleScan).toBe(true);
+    expect(component.show_qr).toBe(true);
+    expect(component.ebsiFlag).toBe(false);
   });
 
-  it('should copy "did-text" to clipboard when copyToClipboard is called with "did-text"', async () => {
-    spyOn(navigator.clipboard, 'writeText');
+  // it('should copy "did-text" to clipboard when copyToClipboard is called with "did-text"', async () => {
+  //  jest.spyOn(navigator.clipboard, 'writeText');
 
-    const didText = 'DID: exampleDid';
-    const expectedClipboardContent = 'exampleDid';
-    document.body.innerHTML = `<div id="did-text">${didText}</div>`;
+  //   const didText = 'DID: exampleDid';
+  //   const expectedClipboardContent = 'exampleDid';
+  //   document.body.innerHTML = `<div id="did-text">${didText}</div>`;
 
-    await component.copyToClipboard('did-text');
+  //   await component.copyToClipboard('did-text');
 
-    expect(navigator.clipboard.writeText).toHaveBeenCalledWith(expectedClipboardContent);
-  });
+  //   expect(navigator.clipboard.writeText).toHaveBeenCalledWith(expectedClipboardContent);
+  // });
 
-  it('should handle error gracefully if clipboard API fails', async () => {
-    spyOn(navigator.clipboard, 'writeText').and.rejectWith('Test error');
-    spyOn(console, 'error');
+  // it('should handle error gracefully if clipboard API fails', async () => {
+  //  jest.spyOn(navigator.clipboard, 'writeText').and.rejectWith('Test error');
+  //  jest.spyOn(console, 'error');
 
-    const didTextContent = 'DID: exampleDid';
-    document.body.innerHTML = `<div id="did-text">${didTextContent}</div>`;
+  //   const didTextContent = 'DID: exampleDid';
+  //   document.body.innerHTML = `<div id="did-text">${didTextContent}</div>`;
 
-    await component.copyToClipboard('did-text');
+  //   await component.copyToClipboard('did-text');
 
-    expect(console.error).toHaveBeenCalledWith('Error al copiar texto al portapapeles:', 'Test error');
-  });
+  //   expect(console.error).toHaveBeenCalledWith('Error al copiar texto al portapapeles:', 'Test error');
+  // });
 
-  it('should generate credential when generateCred is called', () => {
+  it('should generate credential after websocket connection', fakeAsync(() => {
     const mockCredentialOfferUri = 'mockCredentialOfferUri';
-
-    walletServiceSpy.requestCredential.and.returnValue(of());
 
     component.credentialOfferUri = mockCredentialOfferUri;
     component.generateCred();
 
+    // Avanzar el tiempo para simular el retraso en la conexión del WebSocket
+    tick(1000);
+
     expect(walletServiceSpy.requestCredential).toHaveBeenCalledWith(mockCredentialOfferUri);
-  });
+    expect(websocketServiceSpy.connect).toHaveBeenCalled();
+  }));
 
   it('should update the credential list when refresh is called', fakeAsync(() => {
     const mockCredList: VerifiableCredential[] = [
@@ -157,7 +190,7 @@ describe('CredentialsPage', () => {
       }
     ];
 
-    walletServiceSpy.getAllVCs.and.returnValue(of(mockCredList));
+    walletServiceSpy.getAllVCs.mockReturnValue(of(mockCredList));
 
     component.refresh();
     tick();
@@ -214,73 +247,135 @@ describe('CredentialsPage', () => {
       status: CredentialStatus.ISSUED
     };
 
-    walletServiceSpy.deleteVC.and.returnValue(of('Success'));
-    spyOn(component, 'refresh');
+    walletServiceSpy.deleteVC.mockReturnValue(of('Success'));
+    jest.spyOn(component, 'refresh');
     component.vcDelete(testCredential);
     expect(walletServiceSpy.deleteVC).toHaveBeenCalledWith(testCredential.id);
     expect(component.refresh).toHaveBeenCalled();
   });
 
-
   it('ngOnInit should initialize component properties and call refresh', () => {
-    spyOn(component, 'refresh');
+    walletServiceSpy.getAllVCs.mockReturnValue(of([]));
+    jest.spyOn(component, 'refresh');
     component.ngOnInit();
-    expect(component.scaned_cred).toBeFalse();
+    expect(component.scaned_cred).toBe(false);
     expect(component.refresh).toHaveBeenCalled();
   });
 
-  it('qrCodeEmit should process QR code and potentially change state or call services', fakeAsync(() => {
-    spyOn(router, 'navigate');
+  it('qrCodeEmit should process QR code after websocket connection', fakeAsync(() => {
+    jest.spyOn(mockRouter, 'navigate');
     const testQrCode = "someTestQrCode";
-    component.qrCodeEmit(testQrCode);
-    tick();
 
-    expect(router.navigate).toHaveBeenCalledWith(['/tabs/vc-selector/'], { queryParams: { executionResponse: JSON.stringify({}) } });
+    // Simulamos la conexión del WebSocket.
+    component.qrCodeEmit(testQrCode);
+
+    // Avanzar el tiempo para simular el retraso.
+    tick(1000);
+
+    // Ahora comprobamos que el contenido se ejecutó después de la conexión.
+    expect(walletServiceSpy.executeContent).toHaveBeenCalledWith(testQrCode);
+    expect(websocketServiceSpy.connect).toHaveBeenCalled();
+    expect(mockRouter.navigate).toHaveBeenCalledWith(['/tabs/vc-selector/'], { queryParams: { executionResponse: JSON.stringify({}) } });
   }));
 
+
   it('should handle alert Cancel correctly', fakeAsync(() => {
-    let alertController = TestBed.inject(AlertController);
-    spyOn(alertController, 'create').and.returnValue(Promise.resolve({
+    const alertController = TestBed.inject(AlertController);
+    jest.spyOn(alertController, 'create').mockResolvedValue({
       present: () => Promise.resolve(),
       buttons: [
-        {
-          text: 'Cancelar',
-          role: 'cancel',
-          handler: () => {}
-        },
-        {
-          text: 'Aceptar',
-          handler: () => {}
-        }
+        { text: 'Cancelar', role: 'cancel', handler: () => {} },
+        { text: 'Aceptar', handler: () => {} }
       ]
-    } as any));
+    } as any);
 
     component.credentialClick();
     tick();
   }));
 
   it('should handle alert Accept correctly', fakeAsync(() => {
-    let alertController = TestBed.inject(AlertController);
-    spyOn(alertController, 'create').and.returnValue(Promise.resolve({
+    const alertController = TestBed.inject(AlertController);
+    jest.spyOn(alertController, 'create').mockResolvedValue({
       present: () => Promise.resolve(),
       buttons: [
-        {
-          text: 'Cancelar',
-          role: 'cancel',
-          handler: () => {}
-        },
-        {
-          text: 'Aceptar',
-          handler: () => {
-          }
-        }
+        { text: 'Cancelar', role: 'cancel', handler: () => {} },
+        { text: 'Aceptar', handler: () => {} }
       ]
-    } as any));
+    } as any);
 
     component.credentialClick();
     tick();
   }));
 
 
+  it('should untoggle scan and call detectChanges when navigation is outside /tabs/credentials', fakeAsync(() => {
+    const mockNavigationEndEvent = new NavigationEnd(42, '/tabs/credentials', '/new-url/tabs/other-section');
+    walletServiceSpy.getAllVCs.mockReturnValue(of([]));
 
+    const untoggleScanSpy = jest.spyOn(component, 'untoggleScan');
+    const detectChangesSpy = jest.spyOn(component['cdr'], 'detectChanges');
+
+    mockRouter.events.next(mockNavigationEndEvent);
+
+    expect(untoggleScanSpy).toHaveBeenCalled();
+    expect(detectChangesSpy).toHaveBeenCalled();
+  }));
+
+  it('should not untoggle scan if the destination starts with /tabs/credentials', fakeAsync(() => {
+    const mockNavigationEndEvent = new NavigationEnd(42, '/tabs/credentials', '/tabs/credentials/some-subroute');
+    walletServiceSpy.getAllVCs.mockReturnValue(of([]));
+    const untoggleScanSpy = jest.spyOn(component, 'untoggleScan');
+    const detectChangesSpy = jest.spyOn(component['cdr'], 'detectChanges');
+
+    jest.spyOn(mockRouter.events, 'pipe').mockReturnValue(of(mockNavigationEndEvent));
+
+    expect(untoggleScanSpy).not.toHaveBeenCalled();
+    expect(detectChangesSpy).not.toHaveBeenCalled();
+  }));
+
+  it('should close websocket connection after credential request', fakeAsync(() => {
+    component.credentialOfferUri = 'mockCredentialOfferUri';
+    component.generateCred();
+
+    // Simulamos el retraso para que se complete la conexión y solicitud
+    tick(1000);
+
+    expect(walletServiceSpy.requestCredential).toHaveBeenCalled();
+    expect(websocketServiceSpy.closeConnection).toHaveBeenCalled();
+  }));
+
+  it('should close websocket connection if an error occurs', fakeAsync(() => {
+    jest.spyOn(walletServiceSpy, 'requestCredential').mockReturnValueOnce(throwError(() => new Error('Test error')));
+
+    component.credentialOfferUri = 'mockCredentialOfferUri';
+    component.generateCred();
+
+    // Simulamos el retraso para que se complete la conexión y solicitud
+    tick(1000);
+
+    expect(walletServiceSpy.requestCredential).toHaveBeenCalled();
+    expect(websocketServiceSpy.closeConnection).toHaveBeenCalled();
+  }));
+
+
+  it('should log error to cameraLogsService when executeContent fails', fakeAsync(() => {
+    const mockErrorResponse = {
+      error: {
+        title: 'Test Error Title',
+        message: 'Test Error Message',
+        path: '/test/error/path'
+      }
+    };
+    const errorMessage = `${mockErrorResponse.error.title} . ${mockErrorResponse.error.message} . ${mockErrorResponse.error.path}`;
+
+    jest.spyOn(walletServiceSpy, 'executeContent').mockReturnValueOnce(throwError(() => mockErrorResponse));
+
+    const addCameraLogSpy = jest.spyOn((component as any).cameraLogsService, 'addCameraLog');
+
+    component.qrCodeEmit('someQrCode');
+    tick(2000);
+
+    expect(component.toggleScan).toBe(false);
+    expect(addCameraLogSpy).toHaveBeenCalledWith(new Error(errorMessage), 'httpError');
+  }));
 });
