@@ -1,4 +1,4 @@
-import { Component, inject, Input, OnInit } from '@angular/core';
+import { ChangeDetectorRef, Component, DestroyRef, inject, Input, OnInit } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { AlertController, IonicModule } from '@ionic/angular';
@@ -8,12 +8,17 @@ import { QRCodeModule } from 'angularx-qrcode';
 import { WalletService } from 'src/app/services/wallet.service';
 import { VcViewComponent } from '../../components/vc-view/vc-view.component';
 import { TranslateModule, TranslateService } from '@ngx-translate/core';
-import { Router, ActivatedRoute } from '@angular/router';
+import { Router, ActivatedRoute, NavigationEnd } from '@angular/router';
 import { WebsocketService } from 'src/app/services/websocket.service';
 import { DataService } from 'src/app/services/data.service';
 import { VerifiableCredential } from 'src/app/interfaces/verifiable-credential';
+import { CameraLogsService } from 'src/app/services/camera-logs.service';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { filter } from 'rxjs';
 
 const TIME_IN_MS = 3000;
+
+//TODO don't show creds while scanning, separate scan in another component
 
 @Component({
   selector: 'app-credentials',
@@ -56,8 +61,15 @@ export class CredentialsPage implements OnInit {
   private websocket = inject(WebsocketService);
   private dataService = inject(DataService);
   private route = inject(ActivatedRoute);
+  private destroyRef = inject(DestroyRef);
 
-  public constructor(private alertController: AlertController, public translate: TranslateService,) {
+
+  public constructor(
+    private alertController: AlertController,
+    public translate: TranslateService,
+    private cameraLogsService: CameraLogsService,
+    private cdr: ChangeDetectorRef)
+    {
     this.credOfferEndpoint = window.location.origin + '/tabs/home';
     this.route.queryParams.subscribe((params) => {
       this.toggleScan = params['toggleScan'];
@@ -72,12 +84,26 @@ export class CredentialsPage implements OnInit {
         this.did = data;
       }
     });
+
+    this.router.events
+    .pipe(
+      filter(event => event instanceof NavigationEnd),
+      takeUntilDestroyed()
+    )
+    .subscribe((event: NavigationEnd) => {
+      if (this.route.snapshot.routeConfig?.path==='credentials' && !event.urlAfterRedirects.startsWith('/tabs/credentials')) {
+        this.untoggleScan();
+        this.cdr.detectChanges();
+      }
+    });
+
   }
 
   public ngOnInit() {
     this.scaned_cred = false;
     this.refresh();
-    if (this.credentialOfferUri !== '') {
+    // TODO: Find a better way to handle this
+    if (this.credentialOfferUri !== undefined) {
       this.generateCred();
     }
   }
@@ -87,6 +113,7 @@ export class CredentialsPage implements OnInit {
     this.ebsiFlag = false;
   }
 
+  // TODO: This should be moved to the settings page because this is something recreated to ebsi and this option is enabled in the settings page
   public async copyToClipboard(textToCopy: string) {
     let text = '';
 
@@ -131,53 +158,93 @@ export class CredentialsPage implements OnInit {
   }
 
   public qrCodeEmit(qrCode: string) {
+    console.log('qr code emit')
     this.toggleScan = false;
     this.websocket.connect();
-    this.walletService.executeContent(qrCode).subscribe({
-      next: (executionResponse) => {
-        if (qrCode.includes('credential_offer_uri')) {
-          this.from = 'credential';
-          this.isAlertOpen = true;
-          this.scaned_cred = true;
-          setTimeout(() => {
-            this.isAlertOpen = false;
-            this.scaned_cred = false;
-          }, TIME_IN_MS);
-          this.refresh();
-        } else {
-          this.show_qr = false;
-          this.from = '';
-          this.router.navigate(['/tabs/vc-selector/'], {
-            queryParams: {
-              executionResponse: JSON.stringify(executionResponse),
-            },
-          });
-        }
-      },
 
-      error: (err) => {
-        this.toggleScan = true;
-        console.error(err);
-      },
+    // TODO: Instead of using a delay, we should wait for the websocket connection to be established
+    this.delay(1000).then(() => {
+      //todo don't accept qrs that are not to login or get VC
+      this.walletService.executeContent(qrCode)
+      .pipe(
+        takeUntilDestroyed(this.destroyRef))
+        .subscribe({
+          next: (executionResponse) => {
+            // TODO: Instead of analyzing the qrCode, we should check the response and decide what object we need to show depending on the response
+            if (qrCode.includes('credential_offer_uri')) {
+              this.from = 'credential';
+              this.isAlertOpen = true;
+              this.scaned_cred = true;
+              setTimeout(() => {
+                this.isAlertOpen = false;
+                this.scaned_cred = false;
+              }, TIME_IN_MS);
+              this.refresh();
+            } else {
+              // login from verifier
+              this.show_qr = false;
+              this.from = '';
+              this.router.navigate(['/tabs/vc-selector/'], {
+                queryParams: {
+                  executionResponse: JSON.stringify(executionResponse),
+                },
+              });
+            }
+            this.websocket.closeConnection();
+          },
+          error: (httpErrorResponse) => {
+            this.websocket.closeConnection();
+
+            const httpErr = httpErrorResponse?.error;
+            const message = httpErr?.message || httpErrorResponse?.message || 'No error message';
+            const title = httpErr?.title || httpErrorResponse?.title || '(No title)';
+            const path = httpErr?.path || httpErrorResponse?.path || '(No path)';
+
+            const error = title + ' . ' + message + ' . ' + path;
+            this.cameraLogsService.addCameraLog(new Error(error), 'httpError');
+
+            console.error(httpErrorResponse);
+            setTimeout(()=>{
+              this.router.navigate(['/tabs/home'])
+            }, 1000);
+            
+          },
+        });
     });
   }
 
   public generateCred() {
     this.websocket.connect();
-    this.walletService.requestCredential(this.credentialOfferUri).subscribe({
-      next: () => {
-        this.refresh();
-      },
-      error: (err) => {
-        console.error(err);
-      },
+
+    // Esperar un segundo antes de continuar
+    this.delay(1000).then(() => {
+      this.walletService.requestCredential(this.credentialOfferUri).subscribe({
+        next: () => {
+          this.refresh();
+          this.websocket.closeConnection();
+        },
+        error: (err) => {
+          console.error(err);
+          this.websocket.closeConnection();
+        },
+      });
     });
   }
+
 
   public untoggleScan() {
     this.toggleScan = false;
   }
-
+  public handleButtonKeydown(event: KeyboardEvent, action: string): void {
+    if (event.key === 'Enter' || event.key === ' ') {
+      if (action === 'scan') {
+        this.scan();
+      } else if (action === 'getCredential') {
+        this.credentialClick();
+      }
+      event.preventDefault();
+    }
+  }
   public async credentialClick() {
     const alert = await this.alertController.create({
       header: this.translate.instant('credentials.confirmation'),
@@ -209,5 +276,10 @@ export class CredentialsPage implements OnInit {
     });
 
     await alert.present();
+
   }
+  private delay(ms: number) {
+    return new Promise(resolve => setTimeout(resolve, ms));
+  }
+
 }
