@@ -3,8 +3,10 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { IonicModule } from '@ionic/angular';
 import { Router, RouterModule } from '@angular/router';
-import { TranslateModule } from '@ngx-translate/core';
+import { TranslateModule, TranslateService } from '@ngx-translate/core';
 import { DataService } from 'src/app/services/data.service';
+import { CameraLogsService } from 'src/app/services/camera-logs.service';
+import { environment } from 'src/environments/environment';
 
 @Component({
   selector: 'app-settings',
@@ -23,11 +25,15 @@ import { DataService } from 'src/app/services/data.service';
 export class SettingsPage {
   public userName = '';
   public isAlertOpen = false;
+  public featureLogsEnabled = environment.logs_enabled;
 
   public constructor(
     private router: Router,
-    private dataService: DataService
-  ) {}
+    private dataService: DataService,
+    private cameraLogsService: CameraLogsService,
+    private translate: TranslateService
+  ) {
+  }
 
   public goHomeWithEBSI() {
     this.dataService.getDid().subscribe({
@@ -43,4 +49,16 @@ export class SettingsPage {
   public toggleAlert() {
     this.isAlertOpen = !this.isAlertOpen;
   }
+
+  public async sendCameraLogs() {
+    this.translate.get('mailto_permission_alert').subscribe(async (translatedMsg: string) => {
+      try {
+        alert(translatedMsg)
+        await this.cameraLogsService.fetchCameraLogs();
+        this.cameraLogsService.sendCameraLogs();
+      } catch (error) {
+        console.error('Error sending camera logs:', error);
+      }
+  });
+}
 }
