@@ -3,7 +3,7 @@ import {
   HttpTestingController,
 } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
-import { BehaviorSubject, throwError } from 'rxjs';
+import { BehaviorSubject, of, throwError } from 'rxjs';
 import { DataService } from './data.service';
 import { environment } from 'src/environments/environment';
 
@@ -30,7 +30,7 @@ describe('DataService', () => {
 
   it('listenDid should return BehaviorSubject', () => {
     const did = service.listenDid();
-    expect(did instanceof BehaviorSubject).toBe(true);
+    expect(did).toBeInstanceOf(BehaviorSubject);
     expect(did.getValue()).toBe('');
   });
 
@@ -42,15 +42,15 @@ describe('DataService', () => {
     const req = httpTestingController.expectOne(
       `${environment.server_url}${environment.server_uri.ebsi_did_uri}`
     );
-    expect(req.request.method).toEqual('GET');
+    expect(req.request.method).toBe('GET');
     req.flush(mockDid);
 
     expect(service.listenDid().getValue()).toEqual(mockDid);
   });
 
   it('getDid should handle error', () => {
-    const errorSpy = spyOn(service['http'], 'get').and.returnValue(
-      throwError('Err')
+    const errorSpy = jest.spyOn(service['http'], 'get').mockReturnValue(
+      throwError(() => new Error('Err'))
     );
 
     service.getDid().subscribe({
