@@ -1,15 +1,33 @@
-import { TestBed, fakeAsync, tick } from '@angular/core/testing';
+import { flush, TestBed } from '@angular/core/testing';
 import { ToastServiceHandler } from './toast.service';
 import { ToastController } from '@ionic/angular';
 import { TranslateService, TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { TranslateFakeLoader } from '@ngx-translate/core';
-const TIME_IN_MS = 3000;
+import { fakeAsync, tick } from '@angular/core/testing';
+import { of } from 'rxjs';
+const TIME_IN_MS = 5000;
+
+jest.useFakeTimers();
+
 describe('ToastServiceHandler', () => {
   let service: ToastServiceHandler;
-  let translateService: TranslateService;
-  let toastCtrl: ToastController;
+  let translateService: {get:jest.Mock};
+  let translateSpy: jest.SpyInstance;
+  let toastCtrl: {create:jest.Mock};
+  let alert: {present:jest.Mock, dismiss:jest.Mock}
 
   beforeEach(() => {
+    translateService = {
+      get: jest.fn().mockImplementation((str:string)=>of(str))
+    };
+
+    toastCtrl = {
+      create: jest.fn().mockResolvedValue({
+        present: jest.fn().mockImplementation(()=>Promise.resolve()),
+        dismiss: jest.fn().mockImplementation(()=>Promise.resolve(true)),
+      }),
+    } as any;
+
     TestBed.configureTestingModule({
       imports: [
         TranslateModule.forRoot({
@@ -17,126 +35,264 @@ describe('ToastServiceHandler', () => {
         }),
       ],
       providers: [
-        TranslateService,
-        ToastController,
+        { provide: TranslateService, useValue:translateService },
+        { provide: ToastController, useValue: toastCtrl },
         ToastServiceHandler
       ],
     });
     service = TestBed.inject(ToastServiceHandler);
-    translateService = TestBed.inject(TranslateService);
-    toastCtrl = TestBed.inject(ToastController);
+    translateSpy = jest.spyOn(translateService, 'get');
   });
+
 
   it('should be created', () => {
     expect(service).toBeTruthy();
   });
 
-  it('should display a toast for an error message 6', fakeAsync(() => {
+  it('should format message correctly and translate it', fakeAsync(() => {
+    service.showErrorAlert('Any undefined test message');
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith('errors.default');
 
-    spyOn(toastCtrl, 'create').and.callThrough();
+    service.showErrorAlert("The received QR content cannot be processed");
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith('errors.invalid-qr');
 
-    const errorMessage = "Error processing Verifiable Credential";
-    service.showErrorAlert(errorMessage).subscribe();
+    service.showErrorAlert("There are no credentials available to login");
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith('errors.no-credentials-available');
 
+    service.showErrorAlert('There was a problem processing the QR. It might be invalid or already have been used');
+    tick()
+    expect(translateSpy).toHaveBeenCalledWith("errors.failed-qr-process");
 
-    tick(TIME_IN_MS);
+    service.showErrorAlert("Error while fetching credentialOffer from the issuer");
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith("errors.expired-credentialOffer");
+    
+    service.showErrorAlert("Error while deserializing CredentialOffer");
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith("errors.invalid-credentialOffer");
+    
+    service.showErrorAlert("Error while processing Credential Issuer Metadata from the Issuer");
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith("errors.invalid-issuerMetadata");
+    
+    service.showErrorAlert("Error while fetching  Credential from Issuer");
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith("errors.cannot-get-VC");
 
-    expect(toastCtrl.create).toHaveBeenCalled();
+    service.showErrorAlert("Error processing Verifiable Credential");
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith("errors.cannot-save-VC");
+
+    service.showErrorAlert("Incorrect PIN");
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith("errors.incorrect-pin");
+
+    service.showErrorAlert("Unsigned");
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith("errors.unsigned");
+    
+    service.showErrorAlert("ErrorUnsigned");
+    tick();
+    expect(translateSpy).toHaveBeenCalledWith("errors.Errunsigned");
   }));
-  it('should display a toast for an error message 1', fakeAsync(() => {
-
-    spyOn(toastCtrl, 'create').and.callThrough();
-
+ 
+  it('should create alert for an error message 1', async () => {
+    const toastCtrlSpy = jest.spyOn(toastCtrl, 'create');
     const errorMessage = "The received QR content cannot be processed";
-    service.showErrorAlert(errorMessage).subscribe();
 
+    service.showErrorAlert(errorMessage).subscribe(()=>{});
+  
+    expect(translateSpy).toHaveBeenCalledWith('errors.invalid-qr');
+    expect(toastCtrlSpy).toHaveBeenCalled();
+    expect(toastCtrlSpy).toHaveBeenCalledWith(expect.objectContaining(
+      {
+        [errorMessage]:'errors.invalid-qr'
+      }
+    ));
+    const toast = await toastCtrlSpy.mock.results[0].value;
 
-    tick(TIME_IN_MS);
+    expect(toast.present).toHaveBeenCalled();
+    setTimeout(() => {
+      expect(toast.dismiss).toHaveBeenCalled()
+    }, TIME_IN_MS);
 
-    expect(toastCtrl.create).toHaveBeenCalled();
-  }));
-  it('should display a toast for an error message 2', fakeAsync(() => {
-
-    spyOn(toastCtrl, 'create').and.callThrough();
-
+  });
+ 
+  it('should create alert for an error message 2', async () => {
     const errorMessage = "Error while fetching credentialOffer from the issuer";
-    service.showErrorAlert(errorMessage).subscribe();
+    const toastCtrlSpy = jest.spyOn(toastCtrl, 'create');
 
-
-    tick(TIME_IN_MS);
-
+    service.showErrorAlert(errorMessage).subscribe(()=>{});
+  
+    expect(translateSpy).toHaveBeenCalledWith('errors.expired-credentialOffer');
     expect(toastCtrl.create).toHaveBeenCalled();
-  }));
-    it('should display a toast for an error message 3', fakeAsync(() => {
+    expect(toastCtrl.create).toHaveBeenCalledWith(expect.objectContaining(
+      {
+        [errorMessage]:"errors.expired-credentialOffer"
+      }
+    ));
+    const toast = await toastCtrlSpy.mock.results[0].value;
 
-    spyOn(toastCtrl, 'create').and.callThrough();
+    expect(toast.present).toHaveBeenCalled();
+    setTimeout(() => {
+      expect(toast.dismiss).toHaveBeenCalled()
+    }, TIME_IN_MS);
+  });
 
+  it('should create alert for an error message 3', async () => {
     const errorMessage = "Error while deserializing CredentialOffer";
-    service.showErrorAlert(errorMessage).subscribe();
+    const toastCtrlSpy = jest.spyOn(toastCtrl, 'create');
 
-
-    tick(TIME_IN_MS);
-
+    service.showErrorAlert(errorMessage).subscribe(()=>{});
+  
+    expect(translateSpy).toHaveBeenCalledWith('errors.invalid-credentialOffer');
     expect(toastCtrl.create).toHaveBeenCalled();
-  }));
-  it('should display a toast for an error message 4', fakeAsync(() => {
+    expect(toastCtrl.create).toHaveBeenCalledWith(expect.objectContaining(
+      {
+        [errorMessage]:"errors.invalid-credentialOffer"
+      }
+    ));
 
-    spyOn(toastCtrl, 'create').and.callThrough();
+    const toast = await toastCtrlSpy.mock.results[0].value;
 
+    expect(toast.present).toHaveBeenCalled();
+    setTimeout(() => {
+      expect(toast.dismiss).toHaveBeenCalled()
+    }, TIME_IN_MS);
+  });
+
+  it('should create alert for an error message 4', async () => {
     const errorMessage = "Error while processing Credential Issuer Metadata from the Issuer";
-    service.showErrorAlert(errorMessage).subscribe();
+    const toastCtrlSpy = jest.spyOn(toastCtrl, 'create');
 
-
-    tick(TIME_IN_MS);
-
+    service.showErrorAlert(errorMessage).subscribe(()=>{});
+    
+    expect(translateSpy).toHaveBeenCalledWith('errors.invalid-issuerMetadata');
     expect(toastCtrl.create).toHaveBeenCalled();
-  }));
-  it('should display a toast for an error message 5', fakeAsync(() => {
+    expect(toastCtrl.create).toHaveBeenCalledWith(expect.objectContaining(
+      {
+        [errorMessage]: "errors.invalid-issuerMetadata"
+      }
+    ));
+    
+    const toast = await toastCtrlSpy.mock.results[0].value;
 
-    spyOn(toastCtrl, 'create').and.callThrough();
-
+    expect(toast.present).toHaveBeenCalled();
+    setTimeout(() => {
+      expect(toast.dismiss).toHaveBeenCalled()
+    }, TIME_IN_MS);
+  });
+  
+  it('should create alert for an error message 5', async () => {
     const errorMessage = "Error while fetching  Credential from Issuer";
-    service.showErrorAlert(errorMessage).subscribe();
+    const toastCtrlSpy = jest.spyOn(toastCtrl, 'create');
 
-
-    tick(TIME_IN_MS);
-
+    service.showErrorAlert(errorMessage).subscribe(()=>{});
+    
+    expect(translateSpy).toHaveBeenCalledWith('errors.cannot-get-VC');
     expect(toastCtrl.create).toHaveBeenCalled();
-  }));
-  it('should display a toast for an error message 5', fakeAsync(() => {
+    expect(toastCtrl.create).toHaveBeenCalledWith(expect.objectContaining(
+      {
+        [errorMessage]: "errors.cannot-get-VC"
+      }
+    ));
+    
+    const toast = await toastCtrlSpy.mock.results[0].value;
 
-    spyOn(toastCtrl, 'create').and.callThrough();
+    expect(toast.present).toHaveBeenCalled();
+    setTimeout(() => {
+      expect(toast.dismiss).toHaveBeenCalled()
+    }, TIME_IN_MS);
+  });
+  
+  it('should create alert for an error message 6', async () => {
+    const errorMessage = "Error processing Verifiable Credential";
+    const toastCtrlSpy = jest.spyOn(toastCtrl, 'create');
 
+    service.showErrorAlert(errorMessage).subscribe(()=>{});
+    
+    expect(translateSpy).toHaveBeenCalledWith('errors.cannot-save-VC');
+    expect(toastCtrl.create).toHaveBeenCalled();
+    expect(toastCtrl.create).toHaveBeenCalledWith(expect.objectContaining(
+      {
+        [errorMessage]: "errors.cannot-save-VC"
+      }
+    ));
+    const toast = await toastCtrlSpy.mock.results[0].value;
+
+    expect(toast.present).toHaveBeenCalled();
+    setTimeout(() => {
+      expect(toast.dismiss).toHaveBeenCalled()
+    }, TIME_IN_MS);
+  });
+  
+  it('should create alert for an error message 7', async () => {
     const errorMessage = "Incorrect PIN";
-    service.showErrorAlert(errorMessage).subscribe();
+    const toastCtrlSpy = jest.spyOn(toastCtrl, 'create');
 
-
-    tick(TIME_IN_MS);
-
+    service.showErrorAlert(errorMessage).subscribe(()=>{});
+    
+    expect(translateSpy).toHaveBeenCalledWith('errors.incorrect-pin');
     expect(toastCtrl.create).toHaveBeenCalled();
-  }));
-  it('should display a toast for an error message 6', fakeAsync(() => {
+    expect(toastCtrl.create).toHaveBeenCalledWith(expect.objectContaining(
+      {
+        [errorMessage]: "errors.incorrect-pin"
+      }
+    ));
+   
+    const toast = await toastCtrlSpy.mock.results[0].value;
 
-    spyOn(toastCtrl, 'create').and.callThrough();
-
+    expect(toast.present).toHaveBeenCalled();
+    setTimeout(() => {
+      expect(toast.dismiss).toHaveBeenCalled()
+    }, TIME_IN_MS);
+  });
+  
+  it('should create alert for an error message 8', async () => {
     const errorMessage = "Unsigned";
-    service.showErrorAlert(errorMessage).subscribe();
+    const toastCtrlSpy = jest.spyOn(toastCtrl, 'create');
 
-
-    tick(TIME_IN_MS);
-
+    service.showErrorAlert(errorMessage).subscribe(()=>{});
+    
+    expect(translateSpy).toHaveBeenCalledWith('errors.unsigned');
     expect(toastCtrl.create).toHaveBeenCalled();
-  }));
-  it('should display a toast for an error message 7', fakeAsync(() => {
+    expect(toastCtrl.create).toHaveBeenCalledWith(expect.objectContaining(
+      {
+        [errorMessage]: "errors.unsigned"
+      }
+    ));
+    
+    const toast = await toastCtrlSpy.mock.results[0].value;
 
-    spyOn(toastCtrl, 'create').and.callThrough();
-
+    expect(toast.present).toHaveBeenCalled();
+    setTimeout(() => {
+      expect(toast.dismiss).toHaveBeenCalled()
+    }, TIME_IN_MS);
+  });
+  
+  it('should create alert for an error message 9', async () => {
     const errorMessage = "ErrorUnsigned";
-    service.showErrorAlert(errorMessage).subscribe();
+    const toastCtrlSpy = jest.spyOn(toastCtrl, 'create');
 
-
-    tick(TIME_IN_MS);
-
+    service.showErrorAlert(errorMessage).subscribe(()=>{});
+    
+    expect(translateSpy).toHaveBeenCalledWith('errors.Errunsigned');
     expect(toastCtrl.create).toHaveBeenCalled();
-  }));
+    expect(toastCtrl.create).toHaveBeenCalledWith(expect.objectContaining(
+      {
+        [errorMessage]: "errors.Errunsigned"
+      }
+    ));
+    
+    const toast = await toastCtrlSpy.mock.results[0].value;
+
+    expect(toast.present).toHaveBeenCalled();
+    setTimeout(() => {
+      expect(toast.dismiss).toHaveBeenCalled()
+    }, TIME_IN_MS);
+  });
+  
 });
