@@ -2,22 +2,23 @@
   window["env"] = window["env"] || {};
 
   // Environment variables
-  window["env"]["server_url"] = "${WALLET_API_INTERNAL_DOMAIN}";
-  window["env"]["websocket_url"] = "${WALLET_API_WEBSOCKET_EXTERNAL_DOMAIN}";
-  window["env"]["websocket_uri"] = "${WALLET_API_WEBSOCKET_PATH}";
-  window["env"]["iam_url"] = "${IAM_EXTERNAL_DOMAIN}";
-  window["env"]["iam_uri"] = "${IAM_REALM_PATH}";
-  window["env"]["client_id"] = "${IAM_CLIENT_ID}";
-  window["env"]["scope"] = "${IAM_SCOPE}";
-  window["env"]["grant_type"] = "${IAM_GRANT_TYPE}";
-  window["env"]["execute_content_uri"] = "${WALLET_API_EXECUTE_CONTENT_PATH}";
-  window["env"]["request_credential_uri"] = "${WALLET_API_REQUEST_CREDENTIAL_PATH}";
-  window["env"]["verifiable_presentation_uri"] = "${WALLET_API_VERIFIABLE_PRESENTATION_PATH}";
-  window["env"]["credentials_uri"] = "${WALLET_API_CREDENTIALS_PATH}";
-  window["env"]["credentials_by_id_uri"] = "${WALLET_API_CREDENTIALS_BY_ID_PATH}";
-  window["env"]["credentials_signed_by_id_uri"] = "${WALLET_API_REQUEST_SIGNED_CREDENTIAL_PATH}";
-  window["env"]["users_uri"] = "${WALLET_API_USERS_PATH}";
-  window["env"]["ebsi_did_uri"] = "${WALLET_API_EBSI_PATH}";
-  window["env"]["cbor_uri"] = "${WALLET_API_CBOR_PATH}";
-
+  window["env"]["server_url"] = 'http://localhost:8082';
+  window["env"]["websocket_url"] = 'ws://localhost:8082';
+  window["env"]["websocket_uri"] = '/api/v2/pin';
+  window["env"]["iam_url"] ='http://localhost:7002';
+  window["env"]["iam_uri"] = '/realms/wallet';
+  window["env"]["client_id"] = 'auth-client';
+  window["env"]["scope"] = 'openid profile email offline_access';
+  window["env"]["grant_type"] = 'code';
+  window["env"]["execute_content_uri"] = '/api/v2/execute-content';
+  window["env"]["request_credential_uri"] = '/api/v2/request-credential';
+  window["env"]["verifiable_presentation_uri"] = '/api/v2/verifiable-presentation';
+  window["env"]["credentials_uri"] = '/api/v2/credentials';
+  window["env"]["credentials_by_id_uri"] = '/api/v2/credentials?credentialId=';
+  window["env"]["credentials_signed_by_id_uri"] = '/api/v1/request-signed-credential';
+  window["env"]["users_uri"] = '/api/v2/users';
+  window["env"]["ebsi_did_uri"] = '/api/v2/ebsi-did';
+  window["env"]["cbor_uri"] = '/api/v2/vp/cbor';
+  window["env"]["logs_enabled"] = "false";
+  window["env"]["logs_email"] = "domesupport@in2.es";
 })(this);
